Fix base case and move to top of function
Moved the base case out of the if statement that checks if the first
character in a word matches the last one. The base case is now at the
top of the function.

Alo changed the condition of the base case to check if there is one
character or zero instead of two.